src/prediction/requirements.txt
SQLAlchemy
psycopg2-binary
pandas
joblib
scikit-learn
tensorflow